% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversation_multidyads_utilities.R
\name{calc_sim_cor}
\alias{calc_sim_cor}
\title{Calculate correlation between similarity measures}
\usage{
calc_sim_cor(comparison_df)
}
\arguments{
\item{comparison_df}{A data frame output from compare_sim_meas()}
}
\value{
A correlation matrix
}
\description{
This function calculates the correlation between different similarity measures.
}
\examples{
topic_similarities <- list("1" = c(0.5, 0.6, 0.7), "2" = c(0.4, 0.5, 0.6))
lexical_similarities <- list("1" = c(0.6, 0.7, 0.8), "2" = c(0.5, 0.6, 0.7))
comparison_df <- compare_sim_meas(
  list(topic_similarities, lexical_similarities),
  c("Topic", "Lexical")
)
calc_sim_cor(comparison_df)
print(plot)
}
