% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversation_sequence_utilities.R
\name{norm_sim}
\alias{norm_sim}
\title{Normalize Similarity Scores}
\usage{
norm_sim(similarities)
}
\arguments{
\item{similarities}{A numeric vector of similarity scores}
}
\value{
A numeric vector of normalized similarity scores
}
\description{
Normalize similarity scores
}
\details{
This function normalizes similarity scores to a 0-1 range.
}
\examples{
similarities <- c(0.2, 0.5, 0.8, 1.0, 0.3)
norm_sim(similarities)
print(plot)
}
