% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_utilities.R
\name{plot_sims}
\alias{plot_sims}
\title{Visualize similarity scores}
\usage{
plot_sims(similarities)
}
\arguments{
\item{similarities}{A named list of similarity scores}
}
\value{
A ggplot object
}
\description{
This function creates a bar plot of similarity scores.
}
\examples{
sims <- list(topic = 0.8, lexical = 0.6, semantic = 0.7, structural = 0.9)
plot_sims(sims)
}
