% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run package examples}
\usage{
run_example(example_name)
}
\arguments{
\item{example_name}{Name of the example file to run}
}
\value{
No return value, called for side effects.
}
\description{
Run package examples
}
\examples{
\donttest{
run_example("sequence_multidyads_examples.R")
run_example("main_functions_examples.R")
}

}
