% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{sentiment_similarity}
\alias{sentiment_similarity}
\title{Calculate sentiment similarity between two conversations}
\usage{
sentiment_similarity(conv1, conv2)
}
\arguments{
\item{conv1}{A character string representing the first conversation}

\item{conv2}{A character string representing the second conversation}
}
\value{
A numeric value representing the sentiment similarity
}
\description{
This function calculates the sentiment similarity between two conversations
using the sentimentr package.
}
\examples{
conv1 <- "I love this product! It's amazing and works great."
conv2 <- "This item is okay. It does the job but could be better."
sentiment_similarity(conv1, conv2)
}
